% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparACT-package.R
\docType{package}
\name{nparACT}
\alias{nparACT-package}
\alias{nparACT}
\title{nparACT: Non-parametric analysis of actigraphy data}
\description{
Tools for computing classic non-parametric actigraphy measures,
including interdaily stability (IS), intradaily variability (IV),
relative amplitude (RA), as well as the L5 and M10 values
and their respective start times.
}
\author{
\strong{Maintainer}: Christine Blume \email{christine.blume@unibas.ch} (\href{https://orcid.org/0000-0003-2328-9612}{ORCID})

}
