% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparACT_base.R
\name{nparACT_base}
\alias{nparACT_base}
\title{nparACT_base}
\usage{
nparACT_base(
  data,
  SR,
  cutoff = 1,
  plot = FALSE,
  fulldays = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{Data.frame containing the actigraphy data. Actigraphy files must contain
a first column with a date/time vector, which must be of form YYYY-MM-DD
HH:MM:SS or YYYY/MM/DD HH:MM:SS, i.e. in an unambiguous date format and a
second column with actigraphy values.}

\item{SR}{Sampling rate in Hz.}

\item{cutoff}{Can be used to define a cutoff for actigraphy data. Default value is 1, i.e. all
values are taken into account.}

\item{plot}{If TRUE plots are produced. Default is TRUE.}

\item{fulldays}{If TRUE only data from multiples of 24 hours (i.e. full days) are included while
the rest of the data are discarded. Default is TRUE.}

\item{quiet}{If TRUE will supress all warnings.}
}
\value{
Data frame containing IS, IV and RA values. Besides this, L5 and M10 values are given along with
the respective start times.
}
\description{
Function returns IS, IV, RA, L5, M10 as well as L5 and M10 start times for single actigraphy files.
The function also returns a classic dual day plot for the data, a plot of minutewise averages across
24 hours as well as a plot for hourly averages across 24 hours.
}
\examples{
# example function nparACT
data(sleepstudy)
r <- nparACT_base(sleepstudy, SR = 4/60)
}
