% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparACT_flex_loop.R
\name{nparACT_flex_loop}
\alias{nparACT_flex_loop}
\title{nparACT_flex_loop}
\usage{
nparACT_flex_loop(
  path,
  SR,
  cutoff = 1,
  minutes,
  plot = FALSE,
  fulldays = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{path}{Character string containing path to location of actigraphy files. Actigraphy files
must contain a first column with a date/time vector, which must be of form
YYYY-MM-DD HH:MM:SS or YYYY/MM/DD HH:MM:SS, i.e. in an unambiguous date
format and a second column with numeric actigraphy values. Input files can
either be text or comma separated value files.}

\item{SR}{Sampling rate in Hz.}

\item{cutoff}{Can be used to define a cutoff for actigraphy data. Default value is 1, i.e. all
values are taken into account.}

\item{minutes}{Number of minutes for which the Lflex value should be computed.}

\item{plot}{If TRUE plots are produced. Default is TRUE.}

\item{fulldays}{If TRUE only data from multiples of 24 hours (i.e. full days) are included while
the rest of the data are discarded. Default is TRUE.}

\item{quiet}{If TRUE will supress all warnings.}
}
\value{
Data frame containing IS, IV and RA values. Besides this, L5, M10 and Lflex (for the number of
minutes specified) values are given along with the respective start time.
}
\description{
Function returns IS, IV, RA, L5, M10 as well as L5 and M10 start times for multiple actigraphy files.
Additionally, it also returns the L value (Lflex) and the start time for a specified number of minutes.
The function also returns a classic dual day plot for the data, a plot of minutewise averages across
24 hours as well as a plot for hourly averages across 24 hours.
}
\examples{
# example function nparACT_flex_loop
data(sleepstudy)
name <- "sleepstudy_example"
newdir <- file.path(tempdir(),name)
dir.create(newdir, showWarnings = FALSE)
olddir <- setwd(newdir)
write.table(sleepstudy, file = "sleepstudy.txt", row.names=FALSE, col.names = FALSE)
r <- nparACT_flex_loop(newdir, SR = 4/60, minutes = 435)
setwd(olddir)
}
