% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sleepstudy.R
\docType{data}
\name{sleepstudy}
\alias{sleepstudy}
\title{Sleep Study}
\format{
A data frame with 11520 observations on 2 actigraphy variables.
}
\usage{
data(sleepstudy)
}
\description{
Actigraphy data for a sleep study have been collected with wrist actigraphy.
}
\examples{
data(sleepstudy)
}
\keyword{datasets}
