% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nparACT_base_loop.R
\name{nparACT_base_loop}
\alias{nparACT_base_loop}
\title{nparACT_base_loop}
\usage{
nparACT_base_loop(
  path,
  SR,
  cutoff = 1,
  plot = FALSE,
  fulldays = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{path}{Character string containing path to location of actigraphy files. Actigraphy files
must contain a first column with a date/time vector, which must be of form
YYYY-MM-DD HH:MM:SS or YYYY/MM/DD HH:MM:SS, i.e. in an unambiguous date
format and a second column with actigraphy values. Input files can either be
text or comma separated value files.}

\item{SR}{Sampling rate in Hz.}

\item{cutoff}{Can be used to define a cutoff for actigraphy data. Default value is 1, i.e. all
values are taken into account.}

\item{plot}{If TRUE plots are produced. Default is TRUE.}

\item{fulldays}{If TRUE only data from multiples of 24 hours (i.e. full days) are included while
the rest of the data are discarded. Default is TRUE.}

\item{quiet}{If TRUE will supress all warnings.}
}
\value{
Data frame containing IS, IV and RA values for each file. Besides this, L5 and M10 values are
given along with the respective start times
}
\description{
Function returns IS, IV, RA, L5, M10 as well as L5 and M10 start times for multiple actigraphy
files. The function also returns a plot for hourly grand averages across 24 hours.
}
\examples{
# example function nparACT_loop
data(sleepstudy)
name <- "sleepstudy_example"
newdir <- file.path(tempdir(),name)
dir.create(newdir, showWarnings = FALSE)
olddir <- setwd(newdir)
write.table(sleepstudy, file = "sleepstudy.txt", row.names=FALSE, col.names = FALSE)
r <- nparACT_base_loop(newdir, SR = 4/60)
setwd(olddir)
}
