% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepstudy}
\alias{sleepstudy}
\title{Actigraphy results from a participant.}
\format{
A txt file with 20000 rows and 2 variables:
\describe{
\item{Type}{Time stamp}
\item{Description}{Gives the activity measured with actigrapy (arbitrary number)}
}
}
\usage{
sleepstudy
}
\description{
A dataset containing actimetry results for one participant. Data points have been collected every 15 sec and span approx. 3.5 days.
The variables are as follows:
}
\keyword{datasets}
